Add custom profile for shanand user and reorder custom profiles (#20)
This change was necessary based on some experiments that profiles are matched sequencially.

That way, if you have a custom user profile that changes spark cluster configuration, the latter image profiles that does the same changes will be applied.

The order that must be followed should be: global profile -> image-specific profiles -> user-specific profiles